change to double

+++ src/com/sda/she_likes_java/homework/Main.java
@@ -6,15 +6,15 @@ public class Main {
     public static void main(String[] args) {
         System.out.println("Hello world!");
         Scanner sc;
-        Integer num1;
-        Integer num2;
+        double num1;
+        double num2;
         char operator;
 
         System.out.println("Enter first number:");
         sc = new Scanner(System.in);
-        num1 = sc.nextInt();
+        num1 = sc.nextDouble();
         System.out.println("Enter second number:");
-        num2 = sc.nextInt();
+        num2 = sc.nextDouble();
         System.out.println("Please choose operator +,-,/,*");
         operator = sc.next().charAt(0);
